LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-shorthand-expected.txt


PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim" should set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim initial" should not set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim 2000" should not set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim scroll()" should not set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim view()" should not set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim timeline" should not set the property value
PASS Property animation value '1s linear 1s 2 reverse forwards paused anim'
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-delay
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-direction
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-duration
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-fill-mode
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-iteration-count
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-name
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-play-state
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-timeline assert_equals: animation-timeline should be canonical expected (string) "auto" but got (undefined) undefined
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should set animation-timing-function
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1,\n   1s linear 1s 2 reverse forwards paused anim2,\n   1s linear 1s 2 reverse forwards paused anim3" should not set unrelated longhands
FAIL Animation shorthand can not represent non-initial timelines (specified) assert_equals: expected "" but got "1s anim"
FAIL Animation shorthand can not represent non-initial timelines (computed) assert_equals: expected "" but got "1s ease 0s 1 normal none running anim"
FAIL Animation shorthand can not represent non-initial animation-delay-end (specified) assert_equals: expected "" but got "1s anim"
FAIL Animation shorthand can not represent non-initial animation-delay-end (computed) assert_equals: expected "" but got "1s ease 0s 1 normal none running anim"
FAIL Animation shorthand can not represent non-initial animation-range-start (specified) assert_equals: expected "" but got "1s anim"
FAIL Animation shorthand can not represent non-initial animation-range-start (computed) assert_equals: expected "" but got "1s ease 0s 1 normal none running anim"
FAIL Animation shorthand can not represent non-initial animation-range-end (specified) assert_equals: expected "" but got "1s anim"
FAIL Animation shorthand can not represent non-initial animation-range-end (computed) assert_equals: expected "" but got "1s ease 0s 1 normal none running anim"
